requirement.txt
SQLAlchemy==2.0.43
pyodbc==5.2.0
tzdata==2025.2
funkybob==2023.12.0
python-dotenv==1.1.1
psycopg2==2.9.11
python-dotenv==1.1.1
requests==2.32.5
alembic==1.17.0